+++ stem/descriptor/server_descriptor.py
@@ -443,6 +443,7 @@ class ServerDescriptor(Descriptor):
   :var bool extra_info_cache: **\*** flag if a mirror for extra-info documents
   :var str extra_info_digest: upper-case hex encoded digest of our extra-info document
   :var bool eventdns: flag for evdns backend (deprecated, always unset)
+  :var str ntor_onion_key: base64 key used to encrypt EXTEND in the ntor protocol
   :var list or_addresses: **\*** alternative for our address/or_port
     attributes, each entry is a tuple of the form (address (**str**), port
     (**int**), is_ipv6 (**bool**))
@@ -492,6 +493,7 @@ class ServerDescriptor(Descriptor):
     'extra_info_digest': (None, _parse_extrainfo_digest_line),
     'hidden_service_dir': (None, _parse_hidden_service_dir_line),
     'eventdns': (None, _parse_eventdns_line),
+    'ntor_onion_key': (None, _parse_ntor_onion_key_line),
     'or_addresses': ([], _parse_or_address_line),
 
     'read_history_end': (None, _parse_read_history_line),
@@ -515,6 +517,7 @@ class ServerDescriptor(Descriptor):
     'hidden-service-dir': _parse_hidden_service_dir_line,
     'uptime': _parse_uptime_line,
     'protocols': _parse_protocols_line,
+    'ntor-onion-key': _parse_ntor_onion_key_line,
     'or-address': _parse_or_address_line,
     'read-history': _parse_read_history_line,
     'write-history': _parse_write_history_line,
@@ -675,7 +678,6 @@ class RelayDescriptor(ServerDescriptor):
 
   :var str onion_key: **\*** key used to encrypt EXTEND cells
   :var str onion_key_crosscert: signature generated using the onion_key
-  :var str ntor_onion_key: base64 key used to encrypt EXTEND in the ntor protocol
   :var str ntor_onion_key_crosscert: signature generated using the ntor-onion-key
   :var str ntor_onion_key_crosscert_sign: sign of the corresponding ed25519 public key
   :var str signing_key: **\*** relay's long-term identity key
@@ -696,7 +698,6 @@ class RelayDescriptor(ServerDescriptor):
 
     'onion_key': (None, _parse_onion_key_line),
     'onion_key_crosscert': (None, _parse_onion_key_crosscert_line),
-    'ntor_onion_key': (None, _parse_ntor_onion_key_line),
     'ntor_onion_key_crosscert': (None, _parse_ntor_onion_key_crosscert_line),
     'ntor_onion_key_crosscert_sign': (None, _parse_ntor_onion_key_crosscert_line),
     'signing_key': (None, _parse_signing_key_line),
@@ -709,7 +710,6 @@ class RelayDescriptor(ServerDescriptor):
     'router-sig-ed25519': _parse_router_sig_ed25519_line,
     'onion-key': _parse_onion_key_line,
     'onion-key-crosscert': _parse_onion_key_crosscert_line,
-    'ntor-onion-key': _parse_ntor_onion_key_line,
     'ntor-onion-key-crosscert': _parse_ntor_onion_key_crosscert_line,
     'signing-key': _parse_signing_key_line,
     'router-signature': _parse_router_signature_line,
@@ -777,10 +777,12 @@ class BridgeDescriptor(ServerDescriptor):
   <https://collector.torproject.org/formats.html#bridge-descriptors>`_)
 
   :var str ed25519_certificate_hash: sha256 hash of the original identity-ed25519
-  :var str router_digest_sha256: **todo**, needs clarification
+  :var str router_digest_sha256: sha256 digest of this document
 
   .. versionchanged:: 1.5.0
      Added the ed25519_certificate_hash and router_digest_sha256 attributes.
+     Also added ntor_onion_key (previously this only belonged to unsanitized
+     descriptors).
   """
 
   ATTRIBUTES = dict(ServerDescriptor.ATTRIBUTES, **{